Made a python file wit the constants
In these constants the unit is included and the definitiveness (if that's a word) of the variable. So an O means that this value can be taken as true, so for example the atmospheric constants won't change and the cruise speed has already been defined. An ~ means it's defined but maybe not needed and a ? means it's a guess